requirements.txt
python-telegram-bot==20.1
requests==2.31.0

